fix: move release procedure to CONTRIBUTING.md file
The release workflow failed due to:

> `long_description` has syntax errors in markup and would not be rendered on PyPI.

The only change made to the README.md was the addition of the release procedure. Because the README.md is also used as a package description it does not make sense to include the release procedure in that description. Therefor I moved it to the CONTRIBUTING.md file where it makes more sense.